refactor(web): extract tool run post-sync lifecycle
